move snap text field to independent file
