+++ ui/src/main/kotlin/com/midtrans/sdk/uikit/internal/view/SnapCreditCardDetailListItem.kt
@@ -301,159 +301,6 @@ fun SnapSavedCardRadioGroup(
     }
 }
 
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun SnapTextField(
-    value: TextFieldValue,
-    onValueChange: (TextFieldValue) -> Unit,
-    isFocused: Boolean,
-    onFocusChange: (Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    readOnly: Boolean = false,
-    textStyle: TextStyle = SnapTypography.STYLES.snapTextMediumRegular,
-    hint: String? = null,
-    leadingIcon: @Composable (() -> Unit)? = null,
-    trailingIcon: @Composable (() -> Unit)? = null,
-    isError: Boolean = false,
-    singleLine: Boolean = true,
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    BasicTextField(
-        value = value,
-        onValueChange = onValueChange,
-        modifier = modifier.then(
-            Modifier
-                .border(
-                    1.dp,
-                    when {
-                        isError && !isFocused -> {
-                            SnapColors.getARGBColor(INTERACTIVE_BORDER_SUPPORT)
-                        }
-
-                        isFocused -> SnapColors.getARGBColor(LINK_HOVER)
-                        else -> SnapColors.getARGBColor(INTERACTIVE_BORDER_INPUT)
-                    },
-                    RoundedCornerShape(4.dp)
-                )
-                .background(Color.White, RoundedCornerShape(4.dp))
-                .width(200.dp)
-                .height(39.dp)
-                .padding(top = 8.dp, bottom = 8.dp, start = 12.dp, end = 12.dp)
-                .onFocusChanged { focusState ->
-                    onFocusChange(focusState.isFocused)
-                }),
-        keyboardOptions = keyboardOptions,
-        enabled = enabled,
-        readOnly = readOnly,
-        textStyle = textStyle,
-        visualTransformation = visualTransformation,
-        singleLine = singleLine,
-        decorationBox = {
-            TextFieldDefaults.TextFieldDecorationBox(
-                value = value.text,
-                innerTextField = it,
-                singleLine = singleLine,
-                enabled = true,
-                placeholder = hint?.let {
-                    {
-                        Text(text = it, style = SnapTypography.STYLES.snapTextMediumRegular)
-                    }
-                },
-                visualTransformation = visualTransformation,
-                // same interaction source as the one passed to BasicTextField to read focus state
-                // for text field styling
-
-                interactionSource = interactionSource,
-                leadingIcon = leadingIcon,
-                trailingIcon = trailingIcon,
-                // keep vertical paddings but change the horizontal
-                contentPadding = TextFieldDefaults.textFieldWithoutLabelPadding(
-                    start = 0.dp, end = 0.dp, top = 0.dp, bottom = 0.dp
-                )
-            )
-        }
-    )
-}
-
-
-//@OptIn(ExperimentalMaterialApi::class)
-//@Composable
-//fun SnapTextField(
-//    value: String,
-//    onValueChange: (String) -> Unit,
-//    modifier: Modifier = Modifier,
-//    enabled: Boolean = true,
-//    readOnly: Boolean = false,
-//    textStyle: TextStyle = SnapTypography.STYLES.snapTextMediumRegular,
-//    hint: String? = null,
-//    leadingIcon: @Composable (() -> Unit)? = null,
-//    trailingIcon: @Composable (() -> Unit)? = null,
-//    isError: Boolean = false,
-//    visualTransformation: VisualTransformation = VisualTransformation.None,
-//    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
-//) {
-//    val interactionSource = remember { MutableInteractionSource() }
-//    var isFocused by remember { mutableStateOf(false) }
-//    BasicTextField(
-//        value = value,
-//        onValueChange = onValueChange,
-//        modifier = modifier.then(
-//            Modifier
-//                .border(
-//                    1.dp,
-//                    when {
-//                        isError -> {
-//                            SnapColors.getARGBColor(INTERACTIVE_BORDER_SUPPORT)
-//                        }
-//
-//                        isFocused -> SnapColors.getARGBColor(LINK_HOVER)
-//                        else -> SnapColors.getARGBColor(INTERACTIVE_BORDER_INPUT)
-//                    },
-//                    RoundedCornerShape(4.dp)
-//                )
-//                .background(Color.White, RoundedCornerShape(4.dp))
-//                .width(200.dp)
-//                .height(39.dp)
-//                .padding(top = 8.dp, bottom = 8.dp, start = 12.dp, end = 12.dp)
-//                .onFocusChanged { focusState ->
-//                    isFocused = focusState.isFocused
-//                }),
-//        keyboardOptions = keyboardOptions,
-//        enabled = enabled,
-//        readOnly = readOnly,
-//        textStyle = textStyle,
-//        visualTransformation = visualTransformation,
-//        decorationBox = {
-//            TextFieldDefaults.TextFieldDecorationBox(
-//                value = value,
-//                innerTextField = it,
-//                singleLine = true,
-//                enabled = true,
-//                placeholder = hint?.let {
-//                    {
-//                        Text(text = it, style = SnapTypography.STYLES.snapTextMediumRegular)
-//                    }
-//                },
-//                visualTransformation = visualTransformation,
-//                // same interaction source as the one passed to BasicTextField to read focus state
-//                // for text field styling
-//
-//                interactionSource = interactionSource,
-//                leadingIcon = leadingIcon,
-//                trailingIcon = trailingIcon,
-//                // keep vertical paddings but change the horizontal
-//                contentPadding = TextFieldDefaults.textFieldWithoutLabelPadding(
-//                    start = 0.dp, end = 0.dp, top = 0.dp, bottom = 0.dp
-//                )
-//            )
-//        }
-//    )
-//}
-
-
 data class SavedCreditCardFormData(
     val startIcon: Int?,
     val endIcon: Int,
